+++ tests/Stub/fixtures/phpdoc-fqcn/input.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Castor\CommandBuilder;
+
+interface CommandBuilderInterface
+{
+    /**
+     * @return string|array<string|\Stringable|int>
+     */
+    public function getCommand(): array|string;
+}