+++ tests/Stub/fixtures/phpdoc-method/input.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Castor\Foo;
+
+use Symfony\Component\Console;
+use Symfony\Component\Console\Completion\CompletionInput;
+
+/**
+ * @method Console\Input\Input foo(CompletionInput $input)
+ */
+class Bar
+{
+}