+++ tests/Stub/fixtures/phpdoc-not-fqcn/expected.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Castor\Attribute;
+
+class AsArgument extends \Castor\Attribute\AsCommandArgument
+{
+    /**
+     * @param array<string>|callable(\Symfony\Component\Console\Completion\CompletionInput): array<string> $autocomplete
+     */
+    public function test($autocomplete): void
+    {
+    }
+}