+++ tests/Stub/fixtures/phpdoc-phpstan-syntax/expected.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Castor\Foobar;
+
+/**
+ * @param class-string $class
+ */
+function test(string $class): string
+{
+}